Tabs to top in prefs
added window.set_search_enabled(false) which moves the tabs to the top of the preferences window by disabling the search functionality. This is the standard way to achieve top-positioned tabs in Adwaita preferences windows.
The tabs will now appear at the top of the window instead of in a sidebar, providing a more traditional tabbed interface layout that's easier to navigate.